Agregado de campos al formulario Create

+++ Announce/src/AnnounceEverything/Models/Condition.cs
@@ -5,7 +5,7 @@ namespace AnnounceEverything.Models
 {
     public class Condition
     {
-        public byte Id { get; set; }
+        public int Id { get; set; }
 
         [MaxLength(80)]
         public string Name { get; set; }
